Updated migration
