+++ src/controllers/orders/update.controller.ts
@@ -1,16 +1,17 @@
 //External
 import { Request, Response } from "express";
 import "dotenv/config";
-//Services
-import { getOrderFromPaypal } from "../../services/orders/get.service";
+//Helpers
+import { updateOrderFromPaypal } from "../../services/orders/update.service";
 //Enums
 import { statusCode } from "../../enum/http/status-code";
 //Helpers
 import { validateHeadersAndKeys } from "../../helpers/validations/headers/validateHeadersAndKeys";
-//Const-vars
+//Const
 const statusCodeInternalServerError = statusCode.INTERNAL_SERVER_ERROR;
 const statusCodeBadRequest = statusCode.BAD_REQUEST;
 const statusCodeOk = statusCode.OK;
+//Vars
 let orderData: any;
 let eventHeaders: any;
 let checkEventHeadersAndKeys: any;
@@ -19,13 +20,13 @@ let msgLog: string;
 
 
 /**
- * @description Controller to get a order from paypal api
+ * @description Controller to update a order from paypal api
  * @param {any} req any type
  * @param {any} res any type
  * @returns  an object with order information from paypal apiapi
  * @example
  */
-export const getOrderController = async (req: Request, res: Response) => {
+export const updateOrderController = async (req: Request, res: Response) => {
   try {
     //-- start with validation headers and keys  ---
     eventHeaders = req.headers;
@@ -40,21 +41,21 @@ export const getOrderController = async (req: Request, res: Response) => {
     //-- end with validation headers and keys  ---
 
     //-- start with axios order operation  ---
-    orderData = await getOrderFromPaypal(req);
+    orderData = await updateOrderFromPaypal(req);
     //-- end with axios order operation  ---
 
     switch (orderData) {
       case null:
         return res
           .status(statusCodeInternalServerError)
-          .send({ error: "Could not get a order. Check the credentials or id of the order" });
+          .send({ error: "Could not update a order. Check the credentials or id of the order" });
       case orderData != null:
         return res.status(statusCodeOk).send(orderData);
       default:
         return res.status(statusCodeOk).send(orderData);
     }
   } catch (error) {
-    msgResponse = "ERROR in getOrderController() function controller.";
+    msgResponse = "ERROR in updateOrderController() function controller.";
     msgLog = msgResponse + `Caused by ${error}`;
     console.log(msgLog);
     return res
