+++ src/helpers/axios/request/get.ts
@@ -6,22 +6,25 @@ let msgResponse: string;
 let msgLog: string;
 
 /**
- * @description Function to send a axios post request
+ * @description Function to send a axios get request
  * @param {string} url string type
  * @param {any} data any type
  * @param {any} config any type
  * @returns  an object with the information from request
  * @example
  */
-export const sendPostRequest = async (url: string, data: any, config: any) => {
+export const sendGetRequest = async (url: string, data: any, config: any) => {
   try {
-    axiosResponse = await axios.post(url, data, config);
-
+    if (data == (null || undefined)) {
+      axiosResponse = await axios.get(url, config);
+    } else {
+      axiosResponse = await axios.get(url, data, config);
+    }
     axiosData = axiosResponse != null ? axiosResponse.data : null;
 
     return axiosData;
   } catch (error) {
-    msgResponse = "ERROR in sendPostRequest() helper function.";
+    msgResponse = "ERROR in sendGetRequest() helper function.";
     msgLog = msgResponse + `Caused by ${error}`;
     console.log(msgLog);
     return null;
